MAINT: finalize renaming to q2-virsorter2
