+++ .github/workflows/ci.yml
@@ -9,7 +9,7 @@ jobs:
   ci:
     uses: qiime2/distributions/.github/workflows/lib-ci-dev.yaml@dev
     with:
-      distro: metagenome
+      distro: pathogenome
       additional-reports-path: ./coverage.xml
       additional-reports-name: coverage
