+++ q2_virsorter2/__init__.py
@@ -10,7 +10,3 @@
 from ._version import get_versions
 
 __version__ = get_versions()["version"]
-del get_versions
-
-from . import _version
-__version__ = _version.get_versions()['version']
