+++ q2_virsorter2/tests/data/type/vs2_db/group/dsDNAphage/hallmark-gene.list
