+++ q2_virsorter2/tests/data/type/vs2_db/hmm/pfam/Pfam-A.hmm
