+++ q2_virsorter2/tests/data/type/vs2_db_neg/HMM-neg2.hmm
