+++ q2_virsorter2/tests/data/type/vs2_db_neg/Pfam-A-neg2.tsv
