+++ q2_virsorter2/tests/data/type/vs2_db_neg/hallmark-gene-neg5.list
