+++ q2_virsorter2/tests/test_type.py
@@ -7,7 +7,7 @@
 # ----------------------------------------------------------------------------
 from qiime2.plugin.testing import TestPluginBase
 
-from q2_viromics.types._type import Virsorter2Db
+from q2_virsorter2.types._type import Virsorter2Db
 
 
 class TestVirsorter2DbType(TestPluginBase):
