+++ q2_virsorter2/virsorter2_run.py
@@ -14,8 +14,8 @@ import pandas as pd
 import qiime2
 from q2_types.feature_data import DNAFASTAFormat
 
-from q2_viromics._utils import run_command
-from q2_viromics.types._format import Virsorter2DbDirFmt
+from q2_virsorter2._utils import run_command
+from q2_virsorter2.types._format import Virsorter2DbDirFmt
 
 
 # Create the command to fetch the Virsorter2 database
