+++ setup.cfg
@@ -1,10 +1,10 @@
 [versioneer]
 VCS=git
 style=pep440
-versionfile_source = q2_viromics/_version.py
-versionfile_build = q2_viromics/_version.py
+versionfile_source = q2_virsorter2/_version.py
+versionfile_build = q2_virsorter2/_version.py
 tag_prefix =
-parentdir_prefix = q2-viromics-
+parentdir_prefix = q2-virsorter2-
 
 [flake8]
 max-line-length = 88
