+++ setup.py
@@ -13,7 +13,7 @@ import versioneer
 description = "A QIIME 2 plugin for the identification and analysis of viral sequences."
 
 setup(
-    name="q2-viromics",
+    name="q2-virsorter2",
     version=versioneer.get_version(),
     cmdclass=versioneer.get_cmdclass(),
     license="BSD-3-Clause",
@@ -21,13 +21,13 @@ setup(
     author="Christos Matzoros",
     author_email="christosmatzoros@gmail.com",
     description=description,
-    url="https://github.com/bokulich-lab/q2-viromics",
+    url="https://github.com/bokulich-lab/q2-virsorter2",
     entry_points={
-        "qiime2.plugins": ["q2_viromics=" "q2_viromics" ".plugin_setup:plugin"]
+        "qiime2.plugins": ["q2_virsorter2=" "q2_virsorter2" ".plugin_setup:plugin"]
     },
     package_data={
-        "q2_viromics": ["citations.bib"],
-        "q2_viromics.tests": [
+        "q2_virsorter2": ["citations.bib"],
+        "q2_virsorter2.tests": [
             "data/*",
             "data/*/*",
             "data/*/*/*",
